Separately rank the partial solutions for each connected component.
This per-component ranking selects the best partial solution (or at
least minimizes the set of partial solutions) before composing the
results of the partial solutions into set of solutions. This way, we
don't end up creating a huge number of solutions (i.e., all
permutations of the partial solutions) that we'll then have to compare. 

27% improvement in type-checking time for the standard library.



Swift SVN r10865