Introduce a basic worklist into the constraint solver.
Whenever we bind a type variable or merge two type variables, add
those constraints that could be affected to a worklist. Simplification
continues processing constraints in the worklist until the worklist is
empty.

This change reduces the number of constraints that we visit but can't
simplify by ~13x in the standard library, but this doesn't translate
into a performance win. More investigation is needed here.

Note that the worklist is only in use when we have a global constraint
graph, which isn't enabled by default yet.


Swift SVN r10936