When sema synthesizes a derived class init method that has initialized values, it was emitting the super.init() call before the ivar initializers.  Swap this around to correctly obey our model.
Note that Sema does a weird thing where it creates an empty implicit constructor
body, then later comes back around to fill it in.  I'm pretty sure that this is
unnecessary now, but I'm not confident enough in my sema-foo to detangle it.

This is extensively tested in my next patch.


Swift SVN r11072