[sil-extract] Instead of deleting functions, just make them have no basic blocks (i.e. be external).
This gives us as a lot of brevity and simplicity. I plan on adding a trivial
remove globals pass to sil-extract to clean these up.

The mechanism that I am using to do this is:

1. Split the entry basic block at its entrance, insert an unreachable terminator.
2. Use DCE to remove all of the original basic blocks. Thus the function will
only consist of 1 BB whose only instruction is a terminator.
3. Go through and clear the functions basic block list.

The key thing about this approach is that it leaves all of the heavy lifting to
DCE.

Swift SVN r11347