Record witnesses directly in the protocol conformance.
As part of this, centralize the code where we record witnesses. A
little side effect of this is that we end up deducing the RawType
associated type from the derived fromRaw/toRaw witnesses, rather than
deriving RawType ourselves, which is a bit cleaner. The test changes
are because code completion does not know about default definitions in
protocols, which is covered by <rdar://problem/15708469>.


Swift SVN r11505