Work around a closure-in-REPL regression in ContextualizeClosures.
REPL closures in 'if' context trigger another assertion failure, where ContextualizeClosures seems to be getting reentered for mysterious reasons I can't work out. As a workaround, if a closure already has a parent consistent with the parent that would be assigned and a valid discriminator, give up early.

Swift SVN r11552