IRGen: Lower archetype_method calls in -emit-sil-protocol-witness-methods mode.
archetype_method results no longer need to be thick; we can get the Self binding from the substitution in the apply instruction. When considering the 'self' argument of a witness_method, consider an archetype self type as fulfilling itself; we'll inject the metadata for it when we emit the polymorphic source.

Swift SVN r11564