Introduce entry points to resolve a single witness within a conformance.
This is a step toward allowing witnesses to be lazily computed as
needed, rather than requiring an entire explicit conformance to be
checked as a unit.

The new code isn't directly exercised yet. The FIXME'd if(false) in
ConformanceChecker::checkConformance() can be toggled to test these
new entry points, which changes diagnostics but no meaningful behavior
because we'll still computing conformances as a unit.


Swift SVN r11570