IRGen: Populate witness tables using SILWitnessTables.
In -emit-sil-protocol-witness-tables mode, bypass IRGen's thunk generation and consult the SILWitnessTable for the functions to use as witnesses. To minimize the shock to the system, we still produce the witness tables themselves using the broken old lazy generation system; we only use the SILWitnessTable to feed it witnesses.

Swift SVN r11577