IRGen: Ignore 'var' requirements for now in the WitnessBuilder.
I thought we already ignored these, but it turns out we just never actually instantiated witness tables for the stdlib protocols that declare property requirements. When we eagerly generate witness tables this crashed. Ignore them at the IRGen level for now, so we don't need to regress the stdlib.

Swift SVN r11611