Add a skeleton for swift2objc, a tool to generate headers from Swift modules.
This is primarily intended for testing use, but the general mechanism of
generating headers from Swift ASTs is how we plan to get Swift modules into
Clang. The final implementation will be some kind of Clang plugin that's
invoked when a module can't be loaded. The use of header files keeps us
from having to rev-lock Swift and Clang for as long as possible.

Swift SVN r11970