A type function does not override an instance function and vice versa.
We may still have the same problem for type vs. instance variables, though,
but that's not blocking any interaction with Objective-C, which doesn't have
class properties.

This actually caught a mistake in a SILGen test, where a function was
intended to override an Objective-C method but then didn't. (For those
who are wondering why the retains disappeared, the argument is NS_CONSUMED;
therefore, Objective-C is already passing it at +1.)

<rdar://problem/15766044>

Swift SVN r12097